re stucture foldering and refactor
