+++ go.mod
@@ -1,6 +1,6 @@
-module pkg
+module gin-boilerplate
 
-go 1.16
+go 1.17
 
 require (
 	github.com/gin-gonic/gin v1.7.7
