+++ pkg/database/migration.go
@@ -1,6 +1,8 @@
 package database
 
-import "pkg/src/models"
+import (
+	"gin-boilerplate/models"
+)
 
 //Add list of model add for migrations
 //var migrationModels = []interface{}{&models.Example{}, &model.Example{}, &model.Address{})}
