+++ pkg/helpers/search.go
