+++ pkg/routers/index.go
@@ -11,7 +11,7 @@ func RegisterRoutes(route *gin.Engine, db *gorm.DB) {
 	route.NoRoute(func(ctx *gin.Context) {
 		ctx.JSON(http.StatusNotFound, gin.H{"status": http.StatusNotFound, "message": "Route Not Found"})
 	})
-	route.GET("/health", func(ctx *gin.Context) { ctx.JSON(http.StatusOK, gin.H{"like": "ok"}) })
+	route.GET("/health", func(ctx *gin.Context) { ctx.JSON(http.StatusOK, gin.H{"live": "ok"}) })
 
 	//Add All route
 	TestRoutes(route, db)
